requirements.txt
flask==3.0.3
python-whois==0.8.0
dnspython==2.6.1
requests==2.31.0